refactor(hooks): folder-per-hook structure + shared toGetter
Move each hook into its own folder (useX/useX.svelte.ts + spec + index.ts),
mirroring the component folder convention and VueUse. The barrel re-exports
one line per hook folder; the public import 'sv5ui' is unchanged.

Extract the duplicated getter-normalization into a shared hooks/utils.ts
(toGetter) used by useEventListener, useResizeObserver, useScrollLock and
useMediaQuery. useFormField's test helpers move into its folder; internal
component imports updated to the new hook-folder paths.

No public API or behavior change.
