+++ src/lib/Checkbox/Checkbox.svelte
@@ -9,7 +9,7 @@
     import { checkboxVariants, checkboxDefaults } from './checkbox.variants.js'
     import { getComponentConfig, iconsDefaults } from '../config.js'
     import Icon from '../Icon/Icon.svelte'
-    import { useFormField, useFormFieldEmit } from '../hooks/useFormField.svelte.js'
+    import { useFormField, useFormFieldEmit } from '../hooks/useFormField/index.js'
 
     const config = getComponentConfig('checkbox', checkboxDefaults)
     const icons = getComponentConfig('icons', iconsDefaults)
