+++ src/lib/hooks/index.ts
@@ -1,10 +1,10 @@
-export * from './useMediaQuery.svelte.js'
-export * from './useClipboard.svelte.js'
-export * from './useFormField.svelte.js'
-export * from './useClickOutside.svelte.js'
-export * from './useInfiniteScroll.svelte.js'
-export * from './useEscapeKeydown.svelte.js'
-export * from './useDebounce.svelte.js'
-export * from './useEventListener.svelte.js'
-export * from './useResizeObserver.svelte.js'
-export * from './useScrollLock.svelte.js'
+export * from './useMediaQuery/index.js'
+export * from './useClipboard/index.js'
+export * from './useFormField/index.js'
+export * from './useClickOutside/index.js'
+export * from './useInfiniteScroll/index.js'
+export * from './useEscapeKeydown/index.js'
+export * from './useDebounce/index.js'
+export * from './useEventListener/index.js'
+export * from './useResizeObserver/index.js'
+export * from './useScrollLock/index.js'
