+++ src/lib/hooks/useClickOutside/index.ts
@@ -1,0 +1,1 @@
+export * from './useClickOutside.svelte.js'
