+++ src/lib/hooks/useClickOutside/useClickOutside.svelte.ts
