+++ src/lib/hooks/useClipboard/index.ts
@@ -1,0 +1,1 @@
+export * from './useClipboard.svelte.js'
