+++ src/lib/hooks/useClipboard/useClipboard.svelte.spec.ts
