+++ src/lib/hooks/useClipboard/useClipboard.svelte.ts
