+++ src/lib/hooks/useEventListener/index.ts
@@ -1,0 +1,1 @@
+export * from './useEventListener.svelte.js'
