+++ src/lib/hooks/useInfiniteScroll/index.ts
@@ -1,0 +1,1 @@
+export * from './useInfiniteScroll.svelte.js'
