+++ src/lib/hooks/useMediaQuery/index.ts
@@ -1,0 +1,1 @@
+export * from './useMediaQuery.svelte.js'
