+++ src/lib/hooks/useResizeObserver/index.ts
@@ -1,0 +1,1 @@
+export * from './useResizeObserver.svelte.js'
