+++ src/lib/hooks/useScrollLock/index.ts
@@ -1,0 +1,1 @@
+export * from './useScrollLock.svelte.js'
