+++ src/lib/hooks/utils.ts
@@ -1,0 +1,3 @@
+export function toGetter<T>(value: T | (() => T)): () => T {
+    return typeof value === 'function' ? (value as () => T) : () => value
+}
